django/requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
beautifulsoup4==4.9.3
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
cssselect2==0.3.0
cycler==0.10.0
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
fdb==2.0.1
future==0.18.2
idna==2.10
isort==5.6.4
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.6.1
Markdown==3.3.1
matplotlib==3.3.2
mccabe==0.6.1
mysqlclient
ntlm-auth==1.5.0
numpy==1.19.2
pandas==1.1.3
Pillow==8.0.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
# pyodbc==4.0.30
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
reportlab==3.5.53
requests==2.24.0
requests-ntlm==1.1.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
svglib==1.0.1
tinycss2==1.0.2
toml==0.10.1
urllib3==1.25.10
webencodings==0.5.1
wrapt==1.12.1
